Fixes #610 — Adjusts the tableview cell height to match the default system height and according to Dynamic Type sizes
