+++ piwigo/Upload/Pick Photos Album/LocalAlbumsViewController.swift
@@ -20,10 +20,10 @@ protocol LocalAlbumsSelectorDelegate: NSObjectProtocol {
     func didSelectPhotoAlbum(withId: String)
 }
 
-class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableViewDataSource, LocalAlbumsProviderDelegate {
-
+class LocalAlbumsViewController: UIViewController {
+    
     weak var delegate: LocalAlbumsSelectorDelegate?
-
+    
     @IBOutlet var localAlbumsTableView: UITableView!
     
     var categoryId: Int32 = AlbumVars.shared.defaultCategory
@@ -31,27 +31,27 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
     weak var albumDelegate: AlbumViewControllerDelegate?
     
     // Actions to perform after selection
-    private enum pwgAlbumSelectAction : Int {
+    enum pwgAlbumSelectAction : Int {
         case none
         case presentLocalAlbum
         case setAutoUploadAlbum
     }
-    private var wantedAction: pwgAlbumSelectAction = .none
-
+    var wantedAction: pwgAlbumSelectAction = .none
+    
     private var selectPhotoLibraryItemsButton: UIBarButtonItem?
     private var cancelBarButton: UIBarButtonItem?
-    private var hasImagesInPasteboard: Bool = false
-
-    private let maxNberOfAlbumsInSection = 23
-    private var hasLimitedNberOfAlbums: [LocalAlbumType : Bool] = [.pasteboard   : false,
-                                                                   .localAlbums  : false,
-                                                                   .eventsAlbums : false,
-                                                                   .syncedAlbums : false,
-                                                                   .facesAlbums  : false,
-                                                                   .sharedAlbums : false,
-                                                                   .mediaTypes   : false,
-                                                                   .otherAlbums  : false]
-    private lazy var pasteboardTypes : [String] = {
+    var hasImagesInPasteboard: Bool = false
+    
+    let maxNberOfAlbumsInSection = 23
+    var hasLimitedNberOfAlbums: [LocalAlbumType : Bool] = [.pasteboard   : false,
+                                                           .localAlbums  : false,
+                                                           .eventsAlbums : false,
+                                                           .syncedAlbums : false,
+                                                           .facesAlbums  : false,
+                                                           .sharedAlbums : false,
+                                                           .mediaTypes   : false,
+                                                           .otherAlbums  : false]
+    lazy var pasteboardTypes : [String] = {
         return [UTType.image.identifier, UTType.movie.identifier]
     }()
     
@@ -63,15 +63,15 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
         }
         return context
     }()
-
+    
     
     // MARK: - View Lifecycle
     override func viewDidLoad() {
         super.viewDidLoad()
-
+        
         // Title
         title = NSLocalizedString("localAlbums", comment: "Photo Library")
-
+        
         // Button for selecting Photo Library items (.limited access mode)
         selectPhotoLibraryItemsButton = UIBarButtonItem(barButtonSystemItem: .camera, target: self, action: #selector(selectPhotoLibraryItems))
         
@@ -86,7 +86,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
         // Register app becoming active for updating the pasteboard
         NotificationCenter.default.addObserver(self, selector: #selector(checkPasteboard),
                                                name: UIApplication.didBecomeActiveNotification, object: nil)
-
+        
         // Use the LocalAlbumsProvider to fetch albums data.
         LocalAlbumsProvider.shared.fetchedLocalAlbumsDelegate = self
         LocalAlbumsProvider.shared.includingEmptyAlbums = (categoryId == Int32.min)
@@ -118,7 +118,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
             self.localAlbumsTableView.reloadData()
         }
     }
-
+    
     @objc func selectPhotoLibraryItems() {
         // Proposes to change the Photo Library selection
         PHPhotoLibrary.shared().presentLimitedLibraryPicker(from: self)
@@ -128,20 +128,20 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
     @objc func applyColorPalette() {
         // Background color of the view
         view.backgroundColor = PwgColor.background
-
+        
         // Navigation bar appearance
         navigationController?.navigationBar.configAppearance(withLargeTitles: false)
-
+        
         // Table view
         setTableViewMainHeader()
         localAlbumsTableView?.separatorColor = PwgColor.separator
         localAlbumsTableView?.indicatorStyle = AppVars.shared.isDarkPaletteActive ? .white : .black
         localAlbumsTableView?.reloadData()
     }
-
+    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-
+        
         // Determine what to do after selection
         if let caller = delegate {
             if caller is AutoUploadViewController {
@@ -156,7 +156,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
             // Navigation "Cancel" button and identifier
             navigationItem.setLeftBarButton(cancelBarButton, animated: true)
             navigationController?.navigationBar.accessibilityIdentifier = "LocalAlbumsNav"
-
+            
             // Check if there are photos/videos in the pasteboard
             let testTypes = UIPasteboard.general.contains(pasteboardTypes: pasteboardTypes) ? true : false
             let nberPhotos = UIPasteboard.general.itemSet(withPasteboardTypes: pasteboardTypes)?.count ?? 0
@@ -165,23 +165,23 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
         
         // Set colors, fonts, etc.
         applyColorPalette()
-
+        
         // Hide toolbar when returning from the LocalImages / PasteboardImages views
         navigationController?.isToolbarHidden = true
-
+        
         // Navigation "Select Photo Library items" button
         if PHPhotoLibrary.authorizationStatus(for: .readWrite) == .limited {
             navigationItem.setRightBarButton(selectPhotoLibraryItemsButton, animated: true)
         }
-
+        
         // Register Low Power Mode status
         NotificationCenter.default.addObserver(self, selector: #selector(setTableViewMainHeader),
                                                name: Notification.Name.NSProcessInfoPowerStateDidChange, object: nil)
     }
-
+    
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
         super.viewWillTransition(to: size, with: coordinator)
-
+        
         // Save position of collection view
         if localAlbumsTableView.visibleCells.count > 0,
            let cell = localAlbumsTableView.visibleCells.first {
@@ -190,7 +190,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
                 coordinator.animate(alongsideTransition: { [self] _ in
                     self.setTableViewMainHeader()
                     self.localAlbumsTableView.reloadData()
-
+                    
                     // Scroll to previous position
                     self.localAlbumsTableView.scrollToRow(at: indexPath, at: .middle, animated: true)
                 })
@@ -204,7 +204,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
         // Update title of current scene (iPad only)
         view.window?.windowScene?.title = NSLocalizedString("tabBar_upload", comment: "Upload")
     }
-
+    
     @objc func checkPasteboard() {
         DispatchQueue.main.async { [self] in
             // Don't consider the pasteboard if the cateogry is null.
@@ -220,7 +220,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
             localAlbumsTableView.reloadData()
         }
     }
-
+    
     @objc func quitUpload() {
         switch wantedAction {
         case .setAutoUploadAlbum, .none:
@@ -231,7 +231,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
             dismiss(animated: true)
         }
     }
-
+    
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
         
@@ -245,7 +245,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
            LocalAlbumsProvider.shared.sharedAlbums.isEmpty,
            LocalAlbumsProvider.shared.mediaTypes.isEmpty,
            LocalAlbumsProvider.shared.otherAlbums.isEmpty  {
-                delegate?.didSelectPhotoAlbum(withId: "")
+            delegate?.didSelectPhotoAlbum(withId: "")
         }
     }
     
@@ -253,7 +253,7 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
         // Unregister all observers
         NotificationCenter.default.removeObserver(self)
     }
-
+    
     
     // MARK: - UITableView - Header
     @MainActor
@@ -283,319 +283,12 @@ class LocalAlbumsViewController: UIViewController, UITableViewDelegate, UITableV
             localAlbumsTableView.tableHeaderView = headerView
         }
     }
-
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        // Get title of section
-        let albumType = albumTypeFor(section: section)
-        let title = LocalAlbumsProvider.shared.titleForFooterInSectionOf(albumType: albumType)
-        return TableViewUtilities.shared.heightOfHeader(withTitle: title,
-                                                        width: tableView.frame.size.width)
-    }
-
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        // Get title of section
-        let albumType = albumTypeFor(section: section)
-        let title = LocalAlbumsProvider.shared.titleForHeaderInSectionOf(albumType: albumType)
-        return TableViewUtilities.shared.viewOfHeader(withTitle: title)
-    }
-
-    func tableView(_ tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int) {
-        // First section added for pasteboard?
-        if hasImagesInPasteboard && (section == 0) { return }
-        view.layer.zPosition = 0
-    }
-
-    
-    // MARK: - UITableView - Rows
-    private func albumTypeFor(section: Int) -> LocalAlbumType {
-        // First section added for pasteboard?
-        var activeSection: Int = section
-        if hasImagesInPasteboard {
-            switch section {
-            case 0:
-                return .pasteboard
-            default:
-                activeSection -= 1
-            }
-        }
-
-        var counter: Int = -1
-        counter += LocalAlbumsProvider.shared.localAlbums.isEmpty ? 0 : 1
-        if activeSection == counter { return .localAlbums }
-        counter += LocalAlbumsProvider.shared.eventsAlbums.isEmpty ? 0 : 1
-        if activeSection == counter { return .eventsAlbums }
-        counter += LocalAlbumsProvider.shared.syncedAlbums.isEmpty ? 0 : 1
-        if activeSection == counter { return .syncedAlbums }
-        counter += LocalAlbumsProvider.shared.facesAlbums.isEmpty ? 0 : 1
-        if activeSection == counter { return .facesAlbums }
-        counter += LocalAlbumsProvider.shared.sharedAlbums.isEmpty ? 0 : 1
-        if activeSection == counter { return .sharedAlbums }
-        counter += LocalAlbumsProvider.shared.mediaTypes.isEmpty ? 0 : 1
-        if activeSection == counter { return .mediaTypes }
-        counter += LocalAlbumsProvider.shared.otherAlbums.isEmpty ? 0 : 1
-        return .otherAlbums
-    }
-    
-    func numberOfSections(in tableView: UITableView) -> Int {
-        var count = Int.zero
-
-        // Consider non-empty collections
-        count += LocalAlbumsProvider.shared.localAlbums.isEmpty ? 0 : 1
-        count += LocalAlbumsProvider.shared.eventsAlbums.isEmpty ? 0 : 1
-        count += LocalAlbumsProvider.shared.syncedAlbums.isEmpty ? 0 : 1
-        count += LocalAlbumsProvider.shared.facesAlbums.isEmpty ? 0 : 1
-        count += LocalAlbumsProvider.shared.sharedAlbums.isEmpty ? 0 : 1
-        count += LocalAlbumsProvider.shared.mediaTypes.isEmpty ? 0 : 1
-        count += LocalAlbumsProvider.shared.otherAlbums.isEmpty ? 0 : 1
-        
-        // First section added for pasteboard if necessary
-        return count + (hasImagesInPasteboard ? 1 : 0)
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        let albumType = albumTypeFor(section: section)
-        var nberOfAlbums = 0
-        switch albumType {
-        case .pasteboard:
-            return 1
-        case .localAlbums:
-            nberOfAlbums = LocalAlbumsProvider.shared.localAlbums.count
-        case .eventsAlbums:
-            nberOfAlbums = LocalAlbumsProvider.shared.eventsAlbums.count
-        case .syncedAlbums:
-            nberOfAlbums = LocalAlbumsProvider.shared.syncedAlbums.count
-        case .facesAlbums:
-            nberOfAlbums = LocalAlbumsProvider.shared.facesAlbums.count
-        case .sharedAlbums:
-            nberOfAlbums = LocalAlbumsProvider.shared.sharedAlbums.count
-        case .mediaTypes:
-            nberOfAlbums = LocalAlbumsProvider.shared.mediaTypes.count
-        case .otherAlbums:
-            nberOfAlbums = LocalAlbumsProvider.shared.otherAlbums.count
-        }
-        return hasLimitedNberOfAlbums[albumType]! ? min(maxNberOfAlbumsInSection, nberOfAlbums) + 1 : nberOfAlbums
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        var assetCollection: PHAssetCollection?
-        let albumType = albumTypeFor(section: indexPath.section)
-        let isLimited = hasLimitedNberOfAlbums[albumType]!
-        switch albumType {
-        case .pasteboard:
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "LocalAlbumsNoDatesTableViewCell", for: indexPath) as? LocalAlbumsNoDatesTableViewCell
-            else { preconditionFailure("Could not load a LocalAlbumsNoDatesTableViewCell!") }
-            let title = NSLocalizedString("categoryUpload_pasteboard", comment: "Clipboard")
-            let nberPhotos = UIPasteboard.general.itemSet(withPasteboardTypes: pasteboardTypes)?.count ?? NSNotFound
-            cell.configure(with: title, nberPhotos: Int64(nberPhotos))
-            cell.isAccessibilityElement = true
-            return cell
-        case .localAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.localAlbums[indexPath.row]
-            }
-        case .eventsAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.eventsAlbums[indexPath.row]
-            }
-        case .syncedAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.syncedAlbums[indexPath.row]
-            }
-        case .facesAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.facesAlbums[indexPath.row]
-            }
-        case .sharedAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.sharedAlbums[indexPath.row]
-            }
-        case .mediaTypes:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.mediaTypes[indexPath.row]
-            }
-        case .otherAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.otherAlbums[indexPath.row]
-            }
-        }
-
-        // Display [+] button at the bottom of section presenting a limited number of albums
-        guard let aCollection = assetCollection else {
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "LocalAlbumsMoreTableViewCell", for: indexPath) as? LocalAlbumsMoreTableViewCell
-            else { preconditionFailure("Could not load a LocalAlbumsMoreTableViewCell!") }
-            cell.configure()
-            cell.isAccessibilityElement = true
-            return cell
-        }
-        
-        // Case of an album
-        let title = aCollection.localizedTitle ?? "—> ? <——"
-        let nberPhotos = Int64(aCollection.estimatedAssetCount)
-
-        if let startDate = aCollection.startDate, let endDate = aCollection.endDate {
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "LocalAlbumsTableViewCell", for: indexPath) as? LocalAlbumsTableViewCell
-            else { preconditionFailure("Could not load a LocalAlbumsTableViewCell!") }
-            cell.configure(with: title, nberPhotos: nberPhotos, startDate: startDate, endDate: endDate)
-            cell.accessoryType = wantedAction == .setAutoUploadAlbum ? .none : .disclosureIndicator
-            if aCollection.assetCollectionType == .smartAlbum,
-               aCollection.assetCollectionSubtype == .smartAlbumUserLibrary {
-                cell.accessibilityIdentifier = "Recent"
-            }
-            cell.isAccessibilityElement = true
-            return cell
-        }
-        else {
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "LocalAlbumsNoDatesTableViewCell", for: indexPath) as? LocalAlbumsNoDatesTableViewCell
-            else { preconditionFailure("Could not load a LocalAlbumsNoDatesTableViewCell!") }
-            cell.configure(with: title, nberPhotos: nberPhotos)
-            cell.accessoryType = wantedAction == .setAutoUploadAlbum ? .none : .disclosureIndicator
-            if aCollection.assetCollectionType == .smartAlbum,
-               aCollection.assetCollectionSubtype == .smartAlbumUserLibrary {
-                cell.accessibilityIdentifier = "Recent"
-            }
-            cell.isAccessibilityElement = true
-            return cell
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        var assetCollection: PHAssetCollection?
-        let albumType = albumTypeFor(section: indexPath.section)
-        let isLimited = hasLimitedNberOfAlbums[albumType]!
-        switch albumType {
-        case .pasteboard:
-            return 44.0 + TableViewUtilities.rowExtraHeight
-        case .localAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.localAlbums[indexPath.row]
-            }
-        case .eventsAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.eventsAlbums[indexPath.row]
-            }
-        case .syncedAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.syncedAlbums[indexPath.row]
-            }
-        case .facesAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.facesAlbums[indexPath.row]
-            }
-        case .sharedAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.sharedAlbums[indexPath.row]
-            }
-        case .mediaTypes:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.mediaTypes[indexPath.row]
-            }
-        case .otherAlbums:
-            if !(isLimited && indexPath.row == maxNberOfAlbumsInSection) {
-                assetCollection = LocalAlbumsProvider.shared.otherAlbums[indexPath.row]
-            }
-        }
-
-        // Display [+] button at the bottom of section presenting a limited number of albums
-        guard let aCollection = assetCollection else {
-            return 36.0 + TableViewUtilities.rowExtraHeight
-        }
-        
-        // Case of an album
-        if let _ = aCollection.startDate, let _ = aCollection.endDate {
-            return 53.0 + TableViewUtilities.rowExtraHeight
-        } else {
-            return 44.0 + TableViewUtilities.rowExtraHeight
-        }
-    }
-
-    
-    // MARK: - UITableView - Footer
-    func tableView(_ tableView: UITableView, heightForFooterInSection section: Int) -> CGFloat {
-        // Get footer of section
-        let albumType = albumTypeFor(section: section)
-        let footer = LocalAlbumsProvider.shared.titleForFooterInSectionOf(albumType: albumType)
-        return TableViewUtilities.shared.heightOfFooter(withText: footer,
-                                                        width: tableView.frame.width)
-    }
-
-    func tableView(_ tableView: UITableView, viewForFooterInSection section: Int) -> UIView? {
-        // Get footer of section
-        let albumType = albumTypeFor(section: section)
-        let footer = LocalAlbumsProvider.shared.titleForFooterInSectionOf(albumType: albumType)
-        return TableViewUtilities.shared.viewOfFooter(withText: footer, alignment: .center)
-    }
+}
 
 
-    // MARK: - UITableViewDelegate Methods
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-
-        var assetCollections = [PHAssetCollection]()
-        let albumType = albumTypeFor(section: indexPath.section)
-        let isLimited = hasLimitedNberOfAlbums[albumType]!
-        switch albumType {
-        case .pasteboard:
-            let pasteboardImagesSB = UIStoryboard(name: "PasteboardImagesViewController", bundle: nil)
-            guard let pasteboardImagesVC = pasteboardImagesSB.instantiateViewController(withIdentifier: "PasteboardImagesViewController") as? PasteboardImagesViewController else { return }
-            pasteboardImagesVC.categoryId = categoryId
-            pasteboardImagesVC.categoryCurrentCounter = categoryCurrentCounter
-            pasteboardImagesVC.albumDelegate = albumDelegate
-            pasteboardImagesVC.user = user
-            navigationController?.pushViewController(pasteboardImagesVC, animated: true)
-            return
-        case .localAlbums:
-            assetCollections = LocalAlbumsProvider.shared.localAlbums
-        case .eventsAlbums:
-            assetCollections = LocalAlbumsProvider.shared.eventsAlbums
-        case .syncedAlbums:
-            assetCollections = LocalAlbumsProvider.shared.syncedAlbums
-        case .facesAlbums:
-            assetCollections = LocalAlbumsProvider.shared.facesAlbums
-        case .sharedAlbums:
-            assetCollections = LocalAlbumsProvider.shared.sharedAlbums
-        case .mediaTypes:
-            assetCollections = LocalAlbumsProvider.shared.mediaTypes
-        case .otherAlbums:
-            assetCollections = LocalAlbumsProvider.shared.otherAlbums
-        }
-        
-        // Did tap "expand" button at the bottom of section —> release remaining albums
-        if isLimited && indexPath.row == maxNberOfAlbumsInSection {
-            // Release album list
-            hasLimitedNberOfAlbums[albumType] = false
-            // Add remaining albums
-            let indexPaths: [IndexPath] = Array(maxNberOfAlbumsInSection+1..<assetCollections.count)
-                                                .map { IndexPath(row: $0, section: indexPath.section)}
-            tableView.insertRows(at: indexPaths, with: .automatic)
-            // Replace button
-            tableView.reloadRows(at: [indexPath], with: .automatic)
-            return
-        }
-        
-        // Case of an album
-        let albumID = assetCollections[indexPath.row].localIdentifier
-        let albumName = assetCollections[indexPath.row].localizedTitle ?? NSLocalizedString("categoryUpload_LocalAlbums", comment: "Local Albums")
-        if wantedAction == .setAutoUploadAlbum {
-            // Return the selected album ID
-            delegate?.didSelectPhotoAlbum(withId: albumID)
-            navigationController?.popViewController(animated: true)
-        } else {
-            // Presents local images of the selected album
-            let localImagesSB = UIStoryboard(name: "LocalImagesViewController", bundle: nil)
-            guard let localImagesVC = localImagesSB.instantiateViewController(withIdentifier: "LocalImagesViewController") as? LocalImagesViewController else { return }
-            localImagesVC.categoryId = categoryId
-            localImagesVC.categoryCurrentCounter = categoryCurrentCounter
-            localImagesVC.albumDelegate = albumDelegate
-            localImagesVC.imageCollectionId = albumID
-            localImagesVC.imageCollectionName = albumName
-            localImagesVC.user = user
-            navigationController?.pushViewController(localImagesVC, animated: true)
-        }
-    }
-
-    
-    // MARK: - LocalAlbumsProviderDelegate Methods
-    
+// MARK: - LocalAlbumsProviderDelegate Methods
+extension LocalAlbumsViewController: LocalAlbumsProviderDelegate
+{
     func didChangePhotoLibrary() {
         // Change notifications may be made on a background queue. Re-dispatch to the
         // main queue before updating the UI.
